Extend TracerBuilder to facilitate ITracer interception
Background:
When the active Span changes, I want to be notified and do something
with the current SpanContext. One way I can see to do that is to intercept
calls to ITracer::WithSpan.

However, the only way to get an instance of ITracer is through the TracerFactory.
I am proposing that we add a new option to TracerBuilder where people can specify
a Func<ITracer, ITracer> that if specified is invoked for all returned ITracers.

This will allow me to create a proxy ITracer implementation and intercept every call.

usage is:
```csharp

services.AddOpenTelemetry(builder =>
{
    builder
        .SetTracerConstructionInterceptor(t => new TracerProxy(t))

```