[JaegerExporter] Periodically flush undelivered Spans
A Span will never be delivered if the maximum batch size isn't hit or the
component is not torn down. We need to flush these undelivered Spans.

The timer is enabled if a Span is batched up and not sent and then disabled
the next time the batch is flushed.