Fix 'dotnet test OpenTelemetry.sln' out of Windows
The test requires VS to run but even with mono it fails at run time since it references types that only exist on Windows .NET Framework.

The workaround is to not include the test runner outside of Windows, this shows a warning when the command is run outside of Windows.