Allow to filter requests when collecting ASP.NET Core metrics.
This change updates ASP.NET Core instrumentation to respect the filter, specified in the `AspNetCoreInstrumentationOptions.Filter`.

When using the OpenTelemetry library we already could filter out ASP.NET Core request traces, but not metrics. This forced us to completely disable standard metrics instrumentation, because health checks were polluting the measurements and we could not exclude them.

This change updates the `MeterProviderBuilderExtensions` for ASP.NET Core, to accept the `AspNetCoreInstrumentationOptions` configuration delegate and the `HttpInMetricsListener` to respect the filter option.

Code is very similar to the traces instrumentation with one exception - instead of creating the listener in the extension method and passing it to the instrumentation class, I decided to keep listener initialization in the instrumentation class (`AspNetCodeMetrics`), because it already encapsulated .NET diagnostics `Meter`, which is required by the metrics listener. I didn't want to push that code out of the instrumentation class (and deal with new disposable in the outer scope), so I just updated the instrumentation class to take in options in the constructor and pass them through to the listener.

Lastly, I updated the events emitted by the instrumentation itself to, namely `RequestIsFilteredOut` and `RequestFilterException`, to include the name of the handler that was running the filter. This will allow us to distinguish filter delegates specified for metrics vs traces.

It is worth mentioning that I did not update the ASP.NET (non-Core), as it is not an immediate need for us at this point. There are also few more TODOs left to respect other options. We can work on them separately later, if needed.