feat: generate semconv v1.17.0
Generates SemanticConventions from the v1.17.0 schemas. There is an open
issue in the generator that caused mounting subfolders during the
`docker run` steps to fail due to exception.yaml being absent --
mounting that file explicitly made the error go away but it's not clear
whether that produces the desired code or just runs without an error.
There are at least a few changes in the generated source that don't
appear when in the spec changelog when comparing v1.13.0 to v.1.17.0:
- code.column
- messaging.source.kind
- messaging.source.template