[AOT] Trim compatibility of OpenTelemetryLoggerOptions with configuration binder
Configuration binder is currently (.NET 7) incompatible with trimming and AOT. It's essentially a deserializer which uses reflection to access property setters on the options type.

There's no way to make it verifiably correct, but for simple cases it can be validated externally. OpenTelemetryLoggerOptions currently only contains boolean properties, so there are actually no incompatibilities. See the comment in the code for detailed explanation.

The change suppresses the warnings produced by the usage of configuration binder. It also adds a test which verifies that the options type only contains primitive properties. If there were changes to the options type which would make it incomaptible, the test should fail.