Fix collection output buffer management when its resized
When buffer is too small to fit its initial 85kB size, its resized to fit the output, however the resized buffer is not used in final output. This ultimately results in overflow exception as cursor is correct and follows resized buffer.
I also changed to use the same buffer variable for WriteTargetInfo, it also can resize the buffer.